shell arguemtns and clean up

+++ Kernel/memory.h
@@ -1,3 +1,3 @@
 #pragma once
 
-void *memset(void *src, int val, size_t how_many);
+void *memset(void *src, int val, unsigned int how_many);
